requirements.txt
GitPython
Pillow
accelerate

blendmodes
clean-fid
diskcache
einops
facexlib
fastapi>=0.90.1
gradio==3.41.2
inflection
jsonmerge
kornia
lark
numpy
omegaconf
open-clip-torch

piexif
psutil
pytorch_lightning
requests
resize-right

safetensors
scikit-image>=0.19
tomesd
torch
torchdiffeq
torchsde
transformers==4.30.2
pillow-avif-plugin==1.4.3
rembg